requirements.txt
dash==2.17.1
dash_bootstrap_components==1.6.0
python-dotenv==1.0.1
prophet==1.1.5
numpy==1.23.5
